fix(arch): resolve compilation errors in RegVM integration
【动机】
寄存器虚拟机（RegVM）与主程序集成后，存在多处编译错误，包括语法括号闭合错误、类型转换警告、变量命名冲突和控制流逻辑缺陷，导致项目无法通过 MSVC 的 `/WX` 严格编译。

【原理】
1. Emitter.cpp：修正 `Emitter::emit` 函数内部 for 循环的大括号闭合位置，确保 `while (changed)` 循环位于函数体内部，恢复语法正确性。
2. 显式类型转换：将多处 `size_t` 到 `int` 或 `uint8_t` 的隐式转换改为显式 `static_cast`，消除 `/WX` 下的截断警告。
3. 变量遮蔽：移除 `VM::run` 各 `case` 块中重复声明的 `inst` 和 `c` 变量，直接复用外层循环变量，消除遮蔽警告。
4. 帧变量名修正：将 `GET_GLOBAL` 等指令中误用的 `currentFrame` 替换为正确的 `frame` 变量。
5. `ITER_INIT` 控制流重构：将 `if (method) { ... break; }` 改为 `goto` 或调整逻辑，避免 `break` 跳出 `switch` 导致迭代器状态写入被跳过。

+++ src/regvm/compiler/Emitter.cpp
@@ -102,7 +102,7 @@ static int packArgs(std::vector<uint32_t>& words, const std::vector<IRNode*>& ar
     int base = dynamicSpillBase;
     for (size_t i = 0; i < args.size(); ++i) {
         int reg = ensureReg(args[i], words, chunk, 124);
-        auto move = buildInstAB(OpCode::MOVE, base + i, reg);
+        auto move = buildInstAB(OpCode::MOVE, base + static_cast<int>(i), reg);
         words.insert(words.end(), move.begin(), move.end());
     }
     return base;
@@ -371,7 +371,7 @@ int Emitter::emit(IRGraph* graph, Chunk& chunk) {
                     case IROp::IndexSet:
                     case IROp::SliceSet: {
                         int spillBase = packArgs(inst.words, node->dataInputs, chunk, dynamicSpillBase);
-                        auto set = buildInstABC(node->op == IROp::IndexSet ? OpCode::INDEX_SET : OpCode::SLICE_SET, spillBase, 0, node->payload1);
+                        auto set = buildInstABC(node->op == IROp::IndexSet ? OpCode::INDEX_SET : OpCode::SLICE_SET, spillBase, 0, static_cast<int>(node->payload1));
                         inst.words.insert(inst.words.end(), set.begin(), set.end());
                         break;
                     }
